chore: move property animation file
