+++ src/components/anim/curveAnim/AttributeAnimCurve.ts
@@ -1,4 +1,4 @@
-import { AnimationCurve } from "../../../../math/AnimationCurve";
+import { AnimationCurve } from "../../../math/AnimationCurve";
 
 /**
  * @internal
