+++ src/components/anim/curveAnim/PropertyAnimClip.ts
