+++ src/components/anim/curveAnim/PropertyHelp.ts
